サムネが無かったときの判定を追加

+++ app/getConcafeNavi.py
@@ -7,8 +7,8 @@ import csv
 import myFunc
 
 # 定数的なの
-#URLs = ['http://concafenavi.com/area/tokyo/','http://concafenavi.com/area/osaka/,http://concafenavi.com/area/chiba/','http://concafenavi.com/area/aich','http://concafenavi.com/area/fukuoka','http://concafenavi.com/kanagawa','http://concafenavi.com/area/saitama','http://concafenavi.com/area/tochigi','http://concafenavi.com/area/gumma']
-URLs = ['http://concafenavi.com/area/hokkaido/']
+URLs = ['http://concafenavi.com/area/tokyo/','http://concafenavi.com/area/osaka/,http://concafenavi.com/area/chiba/','http://concafenavi.com/area/aich','http://concafenavi.com/area/fukuoka','http://concafenavi.com/kanagawa','http://concafenavi.com/area/saitama','http://concafenavi.com/area/tochigi','http://concafenavi.com/area/gumma']
+#URLs = ['http://concafenavi.com/area/hokkaido/']
 googleDriveFolderId = "1nLocyxN40simk5Ud34s8-LIkmtmNJ2Fs"
 
 
@@ -78,15 +78,18 @@ with open('./getConcafeNavi' + datetime.datetime.now().strftime('%Y%m%d%H%M%S') 
         
         #1/28追加 サムネ画像を取得
         try: 
-            imageURL = d.find_element_by_xpath("/html/body[@class='t-light']/div[@class='l-wrapper']/main[@class='l-main']/section[@class='content content-page']/div[@class='mb20 mt40']/img").get_attribute("src")
-            imageName = "images/" + name + ".png"
-            myFunc.downloadFile(imageURL, imageName)
-            imageID = myFunc.uploadFile(imageName,googleDriveFolderId)
-            myFunc.removeFile(imageName)
+            if not d.find_elements_by_xpath("/html/body[@class='t-light']/div[@class='l-wrapper']/main[@class='l-main']/section[@class='content content-page']/div[@class='mb20 mt40']/img").get_attribute("src") == []:
+                imageURL = d.find_element_by_xpath("/html/body[@class='t-light']/div[@class='l-wrapper']/main[@class='l-main']/section[@class='content content-page']/div[@class='mb20 mt40']/img").get_attribute("src")
+                imageName = "images/" + name + ".png"
+                myFunc.downloadFile(imageURL, imageName)
+                imageID = myFunc.uploadFile(imageName,googleDriveFolderId)
+                myFunc.removeFile(imageName)
 
-            outList.append(imageURL)
-            outList.append(imageName)
-            outList.append(imageID)
+                outList.append(imageURL)
+                outList.append(imageName)
+                outList.append(imageID)
+            else:
+                print("サムネ画像がページに存在しませんでした。")
         except:
             import traceback
             print(traceback.print_exc())
